Module crossbus::rt

Available on crate feature rt only.

Convenient Runtime for common use

it demonstrates the implementor how to implement a runtime, may be removed in the future or replace it with a new crates or a document.

Going Runtime less doesn’t mean execution without runtime, but more procisely, no built-in runtime, but allow any runtime,

it DOES NOT mean CrossBus runs without runtime. on the contrary, CrossBus allow various runtimes and customized runtime even a bare-bone executor.

for convenience, CrossBus provides three common runtimes

bare tokio-based runtime (with feature tokio), just a few crates, however, no io, time supported (build your own if necessary)

async-std-based runtime (with feature async-std), kind of large and redundant

wasm-bindgen-futures-based runtime (with feature wasm32), only spawn_local supported

Modules

runtime_async_stdasync-std

runtime_tokiotokio

runtime_wasm32wasm32

wasm_timeoutwasm32

Traits

SpawnJoinHandle

trait that the return type of Spawning must implement

Spawning

An abstraction for Actor’s Runtime routine